# book-management-model/target/generated-sources/openapi/src/main/java/com/baeldung/openapi/api/BooksApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.8.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.baeldung.openapi.api;

import java.math.BigDecimal;
import com.baeldung.openapi.model.Book;
import com.baeldung.openapi.model.BookRequest;
import com.baeldung.openapi.model.PageContent;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-04T11:39:18.491115100+02:00[Europe/Bucharest]", comments = "Generator version: 7.8.0")
@Validated
@Tag(name = "books", description = "the books API")
public interface BooksApi {

    default BooksApiDelegate getDelegate() {
        return new BooksApiDelegate() {};
    }

    /**
     * POST /books : Add a new book
     *
     * @param bookRequest  (required)
     * @return Book created (status code 200)
     */
    @Operation(
        operationId = "addBook",
        summary = "Add a new book",
        responses = {
            @ApiResponse(responseCode = "200", description = "Book created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Book.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/books",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Book> addBook(
        @Parameter(name = "BookRequest", description = "", required = true) @Valid @RequestBody BookRequest bookRequest
    ) {
        return getDelegate().addBook(bookRequest);
    }


    /**
     * DELETE /books : Delete Book by ID
     *
     * @param id  (required)
     * @return All books deleted (status code 204)
     */
    @Operation(
        operationId = "deleteBook",
        summary = "Delete Book by ID",
        responses = {
            @ApiResponse(responseCode = "204", description = "All books deleted")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/books"
    )
    
    default ResponseEntity<Void> deleteBook(
        @NotNull @Parameter(name = "id", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "id", required = true) BigDecimal id
    ) {
        return getDelegate().deleteBook(id);
    }


    /**
     * GET /books/{id} : Get a book by ID
     *
     * @param id  (required)
     * @return Book details (status code 200)
     */
    @Operation(
        operationId = "getBookById",
        summary = "Get a book by ID",
        responses = {
            @ApiResponse(responseCode = "200", description = "Book details", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Book.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/books/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Book> getBookById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") BigDecimal id
    ) {
        return getDelegate().getBookById(id);
    }


    /**
     * GET /books : List all books
     *
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 5)
     * @return A paginated list of books (status code 200)
     */
    @Operation(
        operationId = "getBooksPage",
        summary = "List all books",
        responses = {
            @ApiResponse(responseCode = "200", description = "A paginated list of books", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PageContent.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/books",
        produces = { "application/json" }
    )
    
    default ResponseEntity<PageContent> getBooksPage(
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "5") Integer size
    ) {
        return getDelegate().getBooksPage(page, size);
    }


    /**
     * PUT /books : Add a new book
     *
     * @param book  (required)
     * @return Book updated (status code 201)
     */
    @Operation(
        operationId = "updateBook",
        summary = "Add a new book",
        responses = {
            @ApiResponse(responseCode = "201", description = "Book updated")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/books",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> updateBook(
        @Parameter(name = "Book", description = "", required = true) @Valid @RequestBody Book book
    ) {
        return getDelegate().updateBook(book);
    }

}